Refactoring. Fixing some uncle tests
