+++ test/libethereum/BlockchainTestsFiller/bcBlockGasLimitTestFiller.json
@@ -19,7 +19,7 @@
         },
         "expect" : {
             "a94f5374fce5edbc8e2a8697c15331677e6ebf0b" : {
-                "balance" : "10000000000"
+                "balance" : "90000000000000009999894980"
             }
         },
         "pre" : {
@@ -75,6 +75,9 @@
         },
         "expect" : {
             "a94f5374fce5edbc8e2a8697c15331677e6ebf0b" : {
+                "balance" : "89999999999999999998739940"
+            },
+	    "aaaf5374fce5edbc8e2a8697c15331677e6ebf0b" : {
                 "balance" : "10000000000"
             }
         },
@@ -185,7 +188,7 @@
         },
         "expect" : {
             "a94f5374fce5edbc8e2a8697c15331677e6ebf0b" : {
-                "balance" : "10000000000"
+                "balance" : "90000000000000009998634920"
             }
         },
         "pre" : {
