+++ test/libethereum/BlockchainTestsFiller/bcGasPricerTestFiller.json
@@ -19,7 +19,10 @@
         },
         "expect" : {
             "a94f5374fce5edbc8e2a8697c15331677e6ebf0b" : {
-                "balance" : "10000000000"
+                "balance" : "8999999995290239999999999890"
+            },
+	    "a94f5374fce5edbc8e2a8697c15331677e6ebf0b" : {
+                "balance" : "8999999995290239999999999890"
             }
         },
         "pre" : {
@@ -30,7 +33,7 @@
                 "storage": {}
             },
             "aaaf5374fce5edbc8e2a8697c15331677e6ebf0b" : {
-                "balance" : "10000000000",
+                "balance" : "10000000100",
                 "nonce" : "0",
                 "code" : "0x60003551",
                 "storage": {}
