+++ test/libethereum/BlockchainTestsFiller/bcInvalidHeaderTestFiller.json
@@ -1142,7 +1142,7 @@
         },
         "expect" : {
             "095e7baea6a6c7c4c2dfeb977efac326af552d87" : {
-                "balance" : "5000000100"
+                "balance" : "100"
             }
         },
         "pre" : {
