+++ test/libethereum/newblockchain.cpp
@@ -4,15 +4,9 @@
 
 #include <test/TestHelper.h>
 #include <test/BlockChainHelper.h>
-
-#include <libethereum/BlockChain.h>
-#include <libethcore/BasicAuthority.h>
 #include <test/JsonSpiritHeaders.h>
-#include <libethereum/TransactionQueue.h>
-#include <libdevcore/TransientDirectory.h>
-#include <libethereum/Block.h>
-
 #include "test/fuzzTesting/fuzzHelper.h"
+
 #include <libdevcore/FileSystem.h>
 #include <libethcore/Params.h>
 
@@ -21,323 +15,6 @@ using namespace json_spirit;
 using namespace dev;
 using namespace dev::eth;
 namespace dev {  namespace test {
-
-TestTransaction::TestTransaction(json_spirit::mObject const& _o): m_jsonTransaction(_o)
-{
-	ImportTest::importTransaction(_o, m_transaction); //check that json structure is valid
-}
-
-TestBlock::TestBlock(TestBlock const& _original)
-{
-	populateFrom(_original);
-}
-
-TestBlock& TestBlock::operator = (TestBlock const& _original)
-{
-	populateFrom(_original);
-	return *this;
-}
-
-TestBlock::TestBlock(json_spirit::mObject& _blockObj, json_spirit::mObject& _stateObj) //genesis block
-{
-	m_state = State(OverlayDB(State::openDB(m_tempDirState.path(), h256{}, WithExisting::Kill)), BaseState::Empty);
-	ImportTest::importState(_stateObj, m_state);
-	m_state.commit();
-
-	m_blockHeader = constructBlock(_blockObj, m_state.rootHash());
-	recalcBlockHeaderBytes(false);
-}
-
-TestBlock::TestBlock(std::string const& _blockRlp)
-{		
-	m_bytes = importByteArray(_blockRlp);
-
-	RLP root(m_bytes);
-	m_blockHeader.populateFromHeader(root[0], IgnoreSeal);
-
-	m_transactionQueue.clear();
-	m_testTransactions.clear();
-	for (auto const& tr: root[1])
-	{
-		Transaction tx(tr.data(), CheckTransaction::Everything);
-		TestTransaction testTx(tx);
-		m_transactionQueue.import(tx.rlp());
-		m_testTransactions.push_back(testTx);
-	}
-
-	for	(auto const& uRLP: root[2])
-	{
-		BlockHeader uBl;
-		uBl.populateFromHeader(uRLP);
-		TestBlock uncle;
-		uncle.setBlockHeader(uBl, false);
-		m_uncles.push_back(uncle);
-	}
-
-	//copyStateFrom(_original.getState());
-}
-
-void TestBlock::setBlockHeader(json_spirit::mObject& _blockHeader)
-{
-	try
-	{
-		const bytes c_rlpBytesBlockHeader = createBlockRLPFromFields(_blockHeader);
-		const RLP c_blockHeaderRLP(c_rlpBytesBlockHeader);
-		m_blockHeader.populateFromHeader(c_blockHeaderRLP, IgnoreSeal);
-	}
-	catch(...)
-	{
-		TBOOST_ERROR("invalid block header json!");
-	}
-}
-
-void TestBlock::addTransaction(TestTransaction const& _tr)
-{
-	m_testTransactions.push_back(_tr);
-	if (m_transactionQueue.import(_tr.getTransaction().rlp()) != ImportResult::Success)
-		cnote << "Test block failed importing transaction\n";
-}
-
-void TestBlock::addUncle(TestBlock const& _uncle)
-{
-	m_uncles.push_back(_uncle);
-}
-
-void TestBlock::setUncles(vector<TestBlock> const& _uncles)
-{
-	m_uncles.clear();
-	m_uncles = _uncles;
-}
-
-void TestBlock::mine(TestBlockChain const& bc)
-{
-	TestBlock const& genesisBlock = bc.getTestGenesis();
-	OverlayDB const& genesisDB = genesisBlock.getState().db();
-
-	FullBlockChain<Ethash> const& blockchain = bc.getInterface();
-
-	Block block = (blockchain.genesisBlock(genesisDB));	
-	block.setBeneficiary(genesisBlock.getBeneficiary());
-	try
-	{
-		ZeroGasPricer gp;
-		block.sync(blockchain);
-		block.sync(blockchain, m_transactionQueue, gp);
-		dev::eth::mine(block, blockchain);
-		//block.sync(blockchain);
-	}
-	catch (Exception const& _e)
-	{
-		cnote << "block sync or mining did throw an exception: " << diagnostic_information(_e);
-		return;
-	}
-	catch (std::exception const& _e)
-	{
-		cnote << "block sync or mining did throw an exception: " << _e.what();
-		return;
-	}
-
-	m_blockHeader = BlockHeader(block.blockData());
-	copyStateFrom(block.state());
-	recalcBlockHeaderBytes();
-}
-
-void TestBlock::setBlockHeader(Ethash::BlockHeader const& _header, bool _recalculate)
-{
-	m_blockHeader = _header;
-	recalcBlockHeaderBytes(_recalculate);
-}
-
-///Test Block Private
-TestBlock::BlockHeader TestBlock::constructBlock(mObject& _o, h256 const& _stateRoot)
-{
-	BlockHeader ret;
-	try
-	{
-		const bytes c_blockRLP = createBlockRLPFromFields(_o, _stateRoot);
-		const RLP c_bRLP(c_blockRLP);
-		ret.populateFromHeader(c_bRLP, IgnoreSeal);
-	}
-	catch (Exception const& _e)
-	{
-		cnote << "block population did throw an exception: " << diagnostic_information(_e);
-	}
-	catch (std::exception const& _e)
-	{
-		TBOOST_ERROR("Failed block population with Exception: " << _e.what());
-	}
-	catch(...)
-	{
-		TBOOST_ERROR("block population did throw an unknown exception\n");
-	}
-	return ret;
-}
-
-bytes TestBlock::createBlockRLPFromFields(mObject const& _tObj, h256 const& _stateRoot)
-{
-	RLPStream rlpStream;
-	rlpStream.appendList(_tObj.count("hash") > 0 ? (_tObj.size() - 1) : _tObj.size());
-
-	if (_tObj.count("parentHash"))
-		rlpStream << importByteArray(_tObj.at("parentHash").get_str());
-
-	if (_tObj.count("uncleHash"))
-		rlpStream << importByteArray(_tObj.at("uncleHash").get_str());
-
-	if (_tObj.count("coinbase"))
-		rlpStream << importByteArray(_tObj.at("coinbase").get_str());
-
-	if (_stateRoot)
-		rlpStream << _stateRoot;
-	else if (_tObj.count("stateRoot"))
-		rlpStream << importByteArray(_tObj.at("stateRoot").get_str());
-
-	if (_tObj.count("transactionsTrie"))
-		rlpStream << importByteArray(_tObj.at("transactionsTrie").get_str());
-
-	if (_tObj.count("receiptTrie"))
-		rlpStream << importByteArray(_tObj.at("receiptTrie").get_str());
-
-	if (_tObj.count("bloom"))
-		rlpStream << importByteArray(_tObj.at("bloom").get_str());
-
-	if (_tObj.count("difficulty"))
-		rlpStream << bigint(_tObj.at("difficulty").get_str());
-
-	if (_tObj.count("number"))
-		rlpStream << bigint(_tObj.at("number").get_str());
-
-	if (_tObj.count("gasLimit"))
-		rlpStream << bigint(_tObj.at("gasLimit").get_str());
-
-	if (_tObj.count("gasUsed"))
-		rlpStream << bigint(_tObj.at("gasUsed").get_str());
-
-	if (_tObj.count("timestamp"))
-		rlpStream << bigint(_tObj.at("timestamp").get_str());
-
-	if (_tObj.count("extraData"))
-		rlpStream << fromHex(_tObj.at("extraData").get_str());
-
-	if (_tObj.count("mixHash"))
-		rlpStream << importByteArray(_tObj.at("mixHash").get_str());
-
-	if (_tObj.count("nonce"))
-		rlpStream << importByteArray(_tObj.at("nonce").get_str());
-
-	return rlpStream.out();
-}
-
-//Form bytestream of a block with [header transactions uncles]
-void TestBlock::recalcBlockHeaderBytes(bool _recalculate)
-{
-	Transactions txList;
-	for (auto const& txi: m_transactionQueue.topTransactions(std::numeric_limits<unsigned>::max()))
-		txList.push_back(txi);
-	RLPStream txStream;
-	txStream.appendList(txList.size());
-	for (unsigned i = 0; i < txList.size(); ++i)
-	{
-		RLPStream txrlp;
-		txList[i].streamRLP(txrlp);
-		txStream.appendRaw(txrlp.out());
-	}
-
-	RLPStream uncleStream;
-	uncleStream.appendList(m_uncles.size());
-	for (unsigned i = 0; i < m_uncles.size(); ++i)
-	{
-		RLPStream uncleRlp;
-		m_uncles[i].getBlockHeader().streamRLP(uncleRlp);
-		uncleStream.appendRaw(uncleRlp.out());
-	}
-
-	if (_recalculate)
-	{
-		if (m_uncles.size()) // update unclehash in case of invalid uncles
-			m_blockHeader.setSha3Uncles(sha3(uncleStream.out()));
-
-		dev::eth::mine(m_blockHeader);
-		m_blockHeader.noteDirty();
-	}
-
-	RLPStream blHeaderStream;
-	m_blockHeader.streamRLP(blHeaderStream, WithProof);
-
-	RLPStream ret(3);
-	ret.appendRaw(blHeaderStream.out()); //block header
-	ret.appendRaw(txStream.out());		 //transactions
-	ret.appendRaw(uncleStream.out());	 //uncles
-
-	m_blockHeader.verifyInternals(&ret.out());
-	m_bytes = ret.out();
-}
-
-void TestBlock::copyStateFrom(State const& _state)
-{
-	//WEIRD WAY TO COPY STATE AS COPY CONSTRUCTOR FOR STATE NOT IMPLEMENTED CORRECTLY (they would share the same DB)
-	m_tempDirState = TransientDirectory();
-	m_state = State(OverlayDB(State::openDB(m_tempDirState.path(), h256{}, WithExisting::Kill)), BaseState::Empty);
-	json_spirit::mObject obj = fillJsonWithState(_state);
-	ImportTest::importState(obj, m_state);
-}
-
-void TestBlock::populateFrom(TestBlock const& _original)
-{
-	copyStateFrom(_original.getState());
-	m_testTransactions = _original.getTestTransactions();
-	m_transactionQueue.clear();
-	TransactionQueue const& trQueue = _original.getTransactionQueue();
-	for (auto const& txi: trQueue.topTransactions(std::numeric_limits<unsigned>::max()))
-		m_transactionQueue.import(txi.rlp());
-
-	m_uncles = _original.getUncles();
-	m_blockHeader = _original.getBlockHeader();
-	m_bytes = _original.getBytes();
-}
-
-///
-TestBlockChain::TestBlockChain(TestBlock const& _genesisBlock)
-{
-	m_blockChain = std::unique_ptr<FullBlockChainEthash>(
-				new FullBlockChainEthash(_genesisBlock.getBytes(), AccountMap(), m_tempDirBlockchain.path(), WithExisting::Kill));
-	m_genesisBlock = _genesisBlock;
-	m_lastBlock = m_genesisBlock;
-}
-
-void TestBlockChain::addBlock(TestBlock const& _block)
-{	
-	m_blockChain.get()->import(_block.getBytes(), m_genesisBlock.getState().db());
-
-	//Imported and best
-	if (_block.getBytes() == m_blockChain.get()->block())
-		m_lastBlock = _block;
-}
-
-vector<TestBlock> TestBlockChain::syncUncles(vector<TestBlock> const& uncles)
-{
-	BlockQueue uncleBlockQueue;
-	vector<TestBlock> validUncles;
-	FullBlockChain<Ethash>& blockchain = *m_blockChain.get();
-	uncleBlockQueue.setChain(blockchain);
-
-	for (size_t i = 0; i < uncles.size(); i++)
-	{
-		try
-		{
-			uncleBlockQueue.import(&uncles.at(i).getBytes(), false);
-			this_thread::sleep_for(chrono::seconds(1)); // wait until block is verified
-			validUncles.push_back(uncles.at(i));
-		}
-		catch(...)
-		{
-			cnote << "error in importing uncle! This produces an invalid block (May be by purpose for testing).";
-		}
-	}
-
-	blockchain.sync(uncleBlockQueue, m_genesisBlock.getState().db(), (unsigned)4);
-	return validUncles;
-}
 
 //Functions that working with test json
 void compareBlocks(TestBlock const& _a, TestBlock const& _b);
@@ -364,7 +41,7 @@ void doBlockchainTests2(json_spirit::mValue& _v, bool _fillin)
 
 		TestBlock genesisBlock(o["genesisBlockHeader"].get_obj(), o["pre"].get_obj());
 		if (_fillin)
-			genesisBlock.setBlockHeader(genesisBlock.getBlockHeader(), true); //update PoW
+			genesisBlock.setBlockHeader(genesisBlock.getBlockHeader(), RecalcBlockHeader::UpdateAndVerify); //update PoW
 		TestBlockChain trueBc(genesisBlock);
 
 		if (_fillin)
@@ -483,10 +160,12 @@ void doBlockchainTests2(json_spirit::mValue& _v, bool _fillin)
 			for (auto const& bl: o["blocks"].get_array())
 			{				
 				mObject blObj = bl.get_obj();
-				TestBlock blockFromRlp(blObj["rlp"].get_str());
+				TestBlock blockFromRlp;
 				try
 				{
-					trueBc.addBlock(blockFromRlp);
+					TestBlock blRlp(blObj["rlp"].get_str());
+					blockFromRlp = blRlp;
+					trueBc.addBlock(blRlp);
 				}
 				// if exception is thrown, RLP is invalid and no blockHeader, Transaction list, or Uncle list should be given
 				catch (Exception const& _e)
@@ -512,26 +191,15 @@ void doBlockchainTests2(json_spirit::mValue& _v, bool _fillin)
 				BOOST_REQUIRE(blObj.count("blockHeader"));
 
 				//Check Provided Header against block in RLP
-				TestBlock blockFromFields;
-				blockFromFields.setBlockHeader(blObj["blockHeader"].get_obj());
+				mObject emptyState;
+				TestBlock blockFromFields(blObj["blockHeader"].get_obj(), emptyState, RecalcBlockHeader::SkipVerify); //not verify as we havent imported transactions and uncles yet
 
 				//ImportTransactions
 				BOOST_REQUIRE(blObj.count("transactions"));
 				for (auto const& txObj: blObj["transactions"].get_array())
 				{
-					try
-					{
-						TestTransaction transaction(txObj.get_obj());
-						blockFromFields.addTransaction(transaction);
-					}
-					catch (Exception const& _e)
-					{
-						TBOOST_ERROR("Failed transaction constructor with Exception: " << diagnostic_information(_e));
-					}
-					catch (exception const& _e)
-					{
-						cnote << _e.what();
-					}
+					TestTransaction transaction(txObj.get_obj());
+					blockFromFields.addTransaction(transaction);
 				}
 
 				// ImportUncles
@@ -541,15 +209,16 @@ void doBlockchainTests2(json_spirit::mValue& _v, bool _fillin)
 						mObject uBlH = uBlHeaderObj.get_obj();
 						BOOST_REQUIRE((uBlH.size() == 16));
 
-						TestBlock uncle;
-						uncle.setBlockHeader(blObj["blockHeader"].get_obj());
+						mObject emptyState;
+						TestBlock uncle(uBlH, emptyState);
 						blockFromFields.addUncle(uncle);
 					}
 
-				checkBlocks(blockFromFields, blockFromRlp, testname);
-				blockFromFields.setBlockHeader(blockFromFields.getBlockHeader(), false); //recalculateBytes
-				try
+				checkBlocks(blockFromFields, blockFromRlp, testname);				
+
+				try				
 				{
+					blockFromFields.setBlockHeader(blockFromFields.getBlockHeader(), RecalcBlockHeader::Verify); //recalculateBytes
 					blockchain.addBlock(blockFromFields);
 				}
 				catch (Exception const& _e)
@@ -564,7 +233,15 @@ void doBlockchainTests2(json_spirit::mValue& _v, bool _fillin)
 			BOOST_CHECK_MESSAGE(lastTrueBlockHash == o["lastblockhash"].get_str(),
 					testname + "Boost check: lastblockhash does not match " + lastTrueBlockHash + " expected: " + o["lastblockhash"].get_str());
 
-			//BOOST_CHECK_MESSAGE(toString(blockchain.getTopBlock().getState().rootHash()) == lastTrueBlockHash, testname + "State root in chain from RLP blocks != State root in chain from Field blocks!");
+			BOOST_CHECK_MESSAGE(toString(trueBc.getTopBlock().getState().rootHash()) ==
+								toString(blockchain.getTopBlock().getState().rootHash()),
+								testname + "State root in chain from RLP blocks != State root in chain from Field blocks!");
+
+			State postState; //Compare post states
+			BOOST_REQUIRE((o.count("postState") > 0));
+			ImportTest::importState(o["postState"].get_obj(), postState);
+			ImportTest::compareStates(postState, trueBc.getTopBlock().getState());
+			ImportTest::compareStates(postState, blockchain.getTopBlock().getState());
 		}
 	}//for tests
 }
@@ -576,7 +253,7 @@ void overwriteBlockHeaderForTest(mObject const& _blObj, TestBlock& _block, Ethas
 	//_block  - which header would be overwritten
 	//_parentHeader - parent blockheader
 
-	bool findNewValidNonce = false;
+	RecalcBlockHeader findNewValidNonce = RecalcBlockHeader::Verify;
 	Ethash::BlockHeader tmp;
 	Ethash::BlockHeader const& header = _block.getBlockHeader();
 	auto ho = _blObj.at("blockHeader").get_obj();
@@ -606,10 +283,13 @@ void overwriteBlockHeaderForTest(mObject const& _blObj, TestBlock& _block, Ethas
 
 		// find new valid nonce
 		if (static_cast<BlockInfo>(tmp) != static_cast<BlockInfo>(header) && tmp.difficulty())
-			findNewValidNonce = true;
+			findNewValidNonce = RecalcBlockHeader::Update;
 
 		if (ho.count("updatePoW"))
-			findNewValidNonce = true;
+			findNewValidNonce = RecalcBlockHeader::UpdateAndVerify;
+
+		tmp.setNonce(header.nonce());
+		tmp.setMixHash(header.mixHash());
 
 		if (ho.count("mixHash"))
 			updateEthashSeal(tmp, h256(ho["mixHash"].get_str()), tmp.nonce());
@@ -622,14 +302,12 @@ void overwriteBlockHeaderForTest(mObject const& _blObj, TestBlock& _block, Ethas
 	{
 		// take the blockheader as is
 		mObject emptyState;
-		tmp = TestBlock(ho, emptyState).getBlockHeader();
+		tmp = TestBlock(ho, emptyState, RecalcBlockHeader::SkipVerify).getBlockHeader();
 	}
 
 	if (_blObj.at("blockHeader").get_obj().count("bruncle"))
 		tmp.populateFromParent(_parentHeader); //importedBlocks.at(importedBlocks.size() -1).getBlockHeader()
 
-	tmp.setNonce(header.nonce());
-	tmp.setMixHash(header.mixHash());
 	_block.setBlockHeader(tmp, findNewValidNonce);
 }
 
@@ -672,10 +350,10 @@ void overwriteUncleHeaderForTest(mObject& uncleHeaderObj, TestBlock& uncle, std:
 			if (prevBlockUncles.size())
 				uncle = prevBlockUncles[0];  //exact uncle??
 			else
-				cerr << "Could not create uncle sameAsPreviousBlockUncle: previous block has no uncles!";
+				cerr << "Could not create uncle sameAsPreviousBlockUncle: previous block has no uncles!" << TestOutputHelper::testName() << endl;
 		}
 		else
-			cerr << "Could not create uncle sameAsPreviousBlockUncle: there are no block imported!";
+			cerr << "Could not create uncle sameAsPreviousBlockUncle: there are no block imported!" << TestOutputHelper::testName() << endl;
 		return;
 	}
 
@@ -705,12 +383,12 @@ void overwriteUncleHeaderForTest(mObject& uncleHeaderObj, TestBlock& uncle, std:
 								 );
 		}
 		else
-			cerr << "Could not create uncle populateFromBlock: there are no block with number " << number;
+			cerr << "Could not create uncle populateFromBlock: there are no block with number " << number << TestOutputHelper::testName() << endl;
 	}
 	else
 	{
 		mObject emptyState;
-		uncle = TestBlock(uncleHeaderObj, emptyState);
+		uncle = TestBlock(uncleHeaderObj, emptyState, RecalcBlockHeader::SkipVerify);
 		uncleHeader = uncle.getBlockHeader();
 	}
 
@@ -719,7 +397,7 @@ void overwriteUncleHeaderForTest(mObject& uncleHeaderObj, TestBlock& uncle, std:
 		uncleHeader = constructHeader(
 			overwrite == "parentHash" ? h256(uncleHeaderObj.at("parentHash").get_str()) : uncleHeader.parentHash(),
 			uncleHeader.sha3Uncles(),
-			uncleHeader.beneficiary(),
+			overwrite == "coinbase" ? Address(uncleHeaderObj.at("coinbase").get_str()) : uncleHeader.beneficiary(),
 			overwrite == "stateRoot" ? h256(uncleHeaderObj.at("stateRoot").get_str()) : uncleHeader.stateRoot(),
 			uncleHeader.transactionsRoot(),
 			uncleHeader.receiptsRoot(),
@@ -737,21 +415,18 @@ void overwriteUncleHeaderForTest(mObject& uncleHeaderObj, TestBlock& uncle, std:
 			uncleHeader.populateFromParent(importedBlocks.at(importedBlocks.size() - 1).getBlockHeader());
 	}
 
-	dev::eth::mine(uncleHeader);
-	uncleHeader.noteDirty();
-	uncle.setBlockHeader(uncleHeader); //updatePoW
-
 	if (overwrite == "nonce" || overwrite == "mixHash")
 	{
-		uncleHeader = uncle.getBlockHeader();
 		if (overwrite == "nonce")
 			updateEthashSeal(uncleHeader, uncleHeader.mixHash(), Nonce(uncleHeaderObj["nonce"].get_str()));
 
 		if (overwrite == "mixHash")
 			updateEthashSeal(uncleHeader, h256(uncleHeaderObj["mixHash"].get_str()), uncleHeader.nonce());
 
-		uncle.setBlockHeader(uncleHeader, false);
+		uncle.setBlockHeader(uncleHeader, RecalcBlockHeader::Verify);
 	}
+	else
+		uncle.setBlockHeader(uncleHeader, RecalcBlockHeader::UpdateAndVerify);
 }
 
 void compareBlocks(TestBlock const& _a, TestBlock const& _b)
@@ -812,8 +487,8 @@ void checkJsonSectionForInvalidBlock(mObject& _blObj)
 
 void checkBlocks(TestBlock const& _blockFromFields, TestBlock const& _blockFromRlp, string const& _testname)
 {
-	Ethash::BlockHeader blockHeaderFromFields = _blockFromFields.getBlockHeader();
-	Ethash::BlockHeader blockFromRlp = _blockFromRlp.getBlockHeader();
+	Ethash::BlockHeader const& blockHeaderFromFields = _blockFromFields.getBlockHeader();
+	Ethash::BlockHeader const& blockFromRlp = _blockFromRlp.getBlockHeader();
 
 	BOOST_CHECK_MESSAGE(blockHeaderFromFields.headerHash(WithProof) == blockFromRlp.headerHash(WithProof), _testname + "hash in given RLP not matching the block hash!");
 	BOOST_CHECK_MESSAGE(blockHeaderFromFields.parentHash() == blockFromRlp.parentHash(), _testname + "parentHash in given RLP not matching the block parentHash!");
@@ -857,12 +532,11 @@ void checkBlocks(TestBlock const& _blockFromFields, TestBlock const& _blockFromR
 		BOOST_CHECK_MESSAGE(trField.rlp() == trRlp.rlp(), _testname + "transactions rlp do not match");
 	}
 
-	BOOST_REQUIRE_EQUAL(_blockFromFields.getUncles().size(), _blockFromRlp.getUncles().size());
-
 	vector<TestBlock> const& unclesFromField = _blockFromFields.getUncles();
 	vector<TestBlock> const& unclesFromRlp = _blockFromRlp.getUncles();
+	BOOST_REQUIRE_EQUAL(unclesFromField.size(), unclesFromRlp.size());
 	for (size_t i = 0; i < unclesFromField.size(); ++i)
-		BOOST_CHECK_MESSAGE(unclesFromField.at(i).getBlockHeader() == unclesFromRlp.at(i).getBlockHeader(), _testname + "block header in rlp and in field do not match");
+		BOOST_CHECK_MESSAGE(unclesFromField.at(i).getBlockHeader() == unclesFromRlp.at(i).getBlockHeader(), _testname + "block header in rlp and in field do not match at uncles");
 }
 
 }}//namespaces
@@ -871,7 +545,7 @@ BOOST_AUTO_TEST_SUITE(BlockChainTests)
 
 BOOST_AUTO_TEST_CASE(bc2)
 {
-	dev::test::executeTests("bcTotalDifficultyTest", "/BlockchainTests",dev::test::getFolder(__FILE__) + "/BlockchainTestsFiller", dev::test::doBlockchainTests2);
+	dev::test::executeTests("bcUncleTest", "/BlockchainTests",dev::test::getFolder(__FILE__) + "/BlockchainTestsFiller", dev::test::doBlockchainTests2);
 }
 
 BOOST_AUTO_TEST_SUITE_END()
